Add the -pkg.el file so the package can be used from Emacs
For Emacs to use the package, the version information has to be specified
explicitly in the Cask file or by the -pkg.el file. We chose to use the -pkg.el
file as we prefer to maintain the information in the main package file
alone (note that the -pkg.el file can be generated by Cask).

This commit also remove the dependency on 'f from the Cask file. It turned out
that it is enough to specify that dependency in the main package file.